Added /isafk <Playername> command
Implemented a command called "/isafk" which takes a Playername as argument and responds with a message stating whether the queried player is marked as AFK or not at the moment.

Also updated readme.md and plugin.yml to include the new command and to increase the version.

Also fixed versioning in plugin.yml (Added a third point for patch-level changes)