Changing the message system
Created the MessageService class, handling all messaging with both players and console - changed all commands and listeners to use the new system, which is also more accessible for other developers

Also added a success message for /rename using the new MessageService messaging system.

To create a MessageService, you have to pass a Map of prefixes and preset messages to the constructor, making them accessible for easy use afterwards. Further information will be provided at a later point in time.